[mt123] Correct Player::SendModifyCooldown

+++ src/game/Player.cpp
@@ -19598,7 +19598,7 @@ void Player::SendModifyCooldown(uint32 spell_id, int32 delta)
     WorldPacket data(SMSG_MODIFY_COOLDOWN, 4 + 8 + 4);
     data << uint32(spell_id);
     data << GetObjectGuid();
-    data << int32(result > 0 ? delta : result - cooldown * IN_MILLISECONDS);
+    data << int32(result > 0 ? delta : result -int32(cooldown) * IN_MILLISECONDS);
     GetSession()->SendPacket(&data);
 }
